pyproject.toml
[project]
name = "hooper-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.0",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "sqlalchemy>=2.0.44",
]
